Fixed Logic Error Tested program and discovered an issue. If you clicked on the format button before selecting a file the program crashed. This is due to a logic error in the code executing changes to the file outside of the safeguards. Moved logic into the File.Exists block where it belongs and added error message in the UI prompting the user to select a file first.
Also took a look at the UI, super simple program so there isn't much to improve on. UI is as simple and intuitive as it gets.

Got rid of extra whitespace in the window.

Color the text window red if program fails (error code posts) and green if it works. Can't get much easier to understand than that.