Lastname, Firstname Convert Fix
Fixed issue with converting to the lastname, firstname format. In the case the user had a middle initial, this format would break and delete / corrupt data. As long as the data provided includes a "." after the middle name (if applicable), this function will now work.

working on a error check to put a "." if the data is missing it.