ToUpper & ToLower Style Implementation Added logic to allow the uppercase and lowercase styles to work. User can now further customize how they want their data changed.
Found and removed duplicate code creating and writing to the file, not sure how that happened..

Fixed weird ui display behavior when user selected display options before selecting a file. When the file was selected the display would blank out causing the preview window to be blank despite the style and format options already being selected.